Working shop and challenges APIs through the api gateway
