+++ apigateway/code/routes/index.js
@@ -1,13 +1,33 @@
 import express from 'express';
 import { createProxyMiddleware } from 'http-proxy-middleware';
-const router = express.Router();
+import cors from 'cors';
 
-// create a proxy for each microservice
+const router = express.Router();
+router.use(cors());
+
 const microserviceProxy = createProxyMiddleware({
   target: 'http://microservice:3011',
   changeOrigin: true
 });
 
 router.use('/microservice', microserviceProxy);
+
+const challengesProxy = createProxyMiddleware({
+  target: 'http://challenges-api:3012',
+  changeOrigin: true,
+});
+
+router.use('/challenges', challengesProxy);
+
+router.get('/health', (req, res) => {
+  res.send('API Gateway is running!');
+});
+
+const itemsProxy = createProxyMiddleware({
+  target: 'http://shop-api:3013',
+  changeOrigin: true,
+});
+
+router.use('/items', itemsProxy);
 
 export default router;
