+++ package.json
@@ -3,6 +3,21 @@
   "dependencies": {
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
-    "express": "^4.21.1"
+    "express": "^4.21.1",
+    "http-proxy": "^1.18.1",
+    "http-proxy-middleware": "^3.0.3"
+  },
+  "name": "uvc-project-backend-team-8-green-quest",
+  "version": "1.0.0",
+  "description": "In this backend example for a project, a folder is created for each micoservice.",
+  "main": "start.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "nodemon": "^3.1.7"
   }
 }
